social_media_api/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
Pillow==11.3.0
sqlparse==0.5.3
gunicorn==23.0.0
python-decouple==3.8
whitenoise==6.8.2
psycopg2-binary==2.9.10
